Fix ArtistFull.Popularity type: int -> float64
Similar artists endpoint returns popularity as 0-1 float (e.g. 0.927)
while search returns it as integer. float64 handles both formats.

+++ internal/types/types.go
@@ -28,7 +28,7 @@ type ArtistFull struct {
 	ID         int    `json:"id"`
 	Name       string `json:"name"`
 	Picture    string `json:"picture"`
-	Popularity int    `json:"popularity"`
+	Popularity float64 `json:"popularity"`
 }
 
 type ArtistSearchResponse struct {
